User/requirements.txt
streamlit
langchain
faiss-cpu
boto3
python-dotenv
langchain_community